Struct Usuario e Credito tiveram o tipo da variável saldo corrigido [float -> double]. O mesmo ocorreu com a variável preco na Struct Produto [float -> double]

+++ escola-de-musica-benicasa/headers.h
@@ -7,7 +7,7 @@ typedef enum {
     ADMINISTRADOR
 } funcao;
 
-struct Usuario{
+struct Usuario {
     int id;                       
     int ativo;                    
     char nome[100];               
@@ -15,7 +15,7 @@ struct Usuario{
     char senha[30];                      
     char especialidade[50];
     funcao categoria;
-    float saldo;
+    double saldo;
 
    Usuario() : id(0), ativo(0), categoria(ALUNO){
         nome[0] = '\0';
@@ -28,7 +28,7 @@ struct Produto {
     char nome[50];
     int id;
     int ativo;
-    float preco;
+    double preco;
     int estoque;
 
     Produto() : id(0), ativo(0), preco(0.0), estoque(0) {
@@ -84,7 +84,7 @@ struct Evento {
 struct Credito {
     int id_opera = 0;
     int id_user = 0;
-    float saldo = 0; 
+    double saldo = 0; 
     bool realizado = false
 };
 #endif
